Adding a couple proposed hacks/work-arounds to the issue of mixed UNICODE and STRING types in our cx_Oracle params.
One is to simply change the reported version number of cx_Oracle to
something below 5.1.

The other is to swap out the convert_unicode function in
django.db.backend.oracle.base within a context manager, so that we can
try to limit the scope of the hack. I don't know of anything else in
Django that looks at cx_Oracle.version, but it may not be good to change
convert_unicode for EVERYTHING.

Added a copy of a failing test that uses the context manager - it passes